requirements.txt
streamlit
PyPDF2
langchain
openai
langchain_openai
pytesseract
pillow
langchain_community
scikit-learn
streamlit-authenticator
